refactor(rho, force): extract "ATOM_ID_TO_LATTICE" and "itl_atoms_pair" in func `itl_atoms_pair` out
In this commit, we refactor `func itl_atoms_pair`, which is used in pairs iteration for calculation
atoms rho and force. In the refactoring, we move the part of "convert lattice id to lattice postion"
and "neighbor iteration for rho/force calculation" out as independent function/macro, which is macro
"ATOM_ID_TO_LATTICE" and inline function "nei_interaction".