fix(compile): move data copying of `d_domain` to the file where define it
When using `hipMemcpyToSymbol` to set device variable, it may cause error of "undefined reference to
`d_domain'", if `hipMemcpyToSymbol` is not called in the same file as the device-side variable
`d_domain` is defined. We can fix it by moving calling of `hipMemcpyToSymbol` to the same file where
define it.

see also: https://github.com/ROCm-Developer-Tools/HIP/issues/2249#issuecomment-810322623